refactor(components): move ThemeToggle to its own folder and update Home
- Moved ThemeToggle.tsx and ThemeToggle.stories.tsx into src/components/themeToggle/
- Updated Home.tsx to reflect the new import path for ThemeToggle
- This change improves code organization and maintainability by grouping related files
